Moved interceptor tests
